refine api
