+++ infra/api/lambda-server/package.json
@@ -5,9 +5,9 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "ts-node src/local.ts",
     "prebuild": "rm -rf dist",
-    "start": "node dist/index.js",
-    "build": "esbuild src/index.ts --bundle --minify --sourcemap --platform=node --target=es2020 --outfile=dist/index.js",
+    "build": "esbuild src/handler.ts --bundle --minify --sourcemap --platform=node --target=es2020 --outfile=dist/index.js",
     "postbuild": "cd dist && zip -r index.zip index.js*"
   },
   "keywords": [],
@@ -28,6 +28,7 @@
   "devDependencies": {
     "@types/aws-lambda": "^8.10.150",
     "esbuild": "^0.25.5",
+    "ts-node": "^10.9.2",
     "typescript": "^5.8.3"
   }
 }
