+++ infra/api/orcavault.tfvars
@@ -1,0 +1,1 @@
+db_name = "orcavault"
